RE-364 Migrate RelluShield to new ItemRegistry
