+++ src/main/java/de/relluem94/minecraft/server/spigot/essentials/constants/ItemConstants.java
@@ -17,6 +17,7 @@ public interface ItemConstants {
 
     String PLUGIN_ITEM_DUMMY = ChatColor.AQUA + "This is a Dummy Item";
 
+    String PLUGIN_ITEM_NAMESPACE_RELLU_SHIELD = "rellu_shield";
     String PLUGIN_ITEM_NAMESPACE_RELLU_HELMET = "rellu_helmet";
     String PLUGIN_ITEM_NAMESPACE_RELLU_CHESTPLATE = "rellu_chestplate";
     String PLUGIN_ITEM_NAMESPACE_RELLU_LEGGINGS = "rellu_leggings";
